+++ src/app/util/transaction.service.ts
@@ -3,13 +3,14 @@ import {Subject} from 'rxjs';
 
 @Injectable()
 export class TxService {
-  public txObservable = new Subject<string[]>();
+  public txObservable = new Subject<string>();
 
   constructor() {
 
   }
 
   updateTx(tx) {
+      console.log(tx);
     this.txObservable.next(tx);
   }
 }